requirements.txt
# Cloud Provider SDKs
boto3>=1.26.0
azure-mgmt-security>=3.0.0
google-cloud-security-center>=1.20.0

# Infrastructure as Code
python-terraform>=0.10.1

# API and Web Framework
fastapi>=0.68.0
uvicorn>=0.15.0

# Database
sqlalchemy>=1.4.0
alembic>=1.7.0

# Monitoring and Alerting
prometheus-client>=0.12.0
python-jose>=3.3.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Utilities
pyyaml>=6.0.0
python-dotenv>=0.19.0
requests>=2.26.0
schedule>=1.1.0
rich>=10.12.0

# Security
cryptography>=37.0.0
passlib>=1.7.4 